drm: xilinx: Add a plane support
This patch adds a plane support of Xilinx DRM KMS driver.
This enables the multi-plane rendering support of the hardware pipeline.
To enable multi-plane support, the hardware composer such as OSD should be
configured properly in the hardware pipeline, and additional dma IPs may
be needed. Otherwise, only the single plane is available,
and a single plane is not exposed to user.

For current implementation, this patch depends on VDMA through dmaengine API,
and an osd sub-driver(xilinx_osd.c/h).

Signed-off-by: Hyun Kwon <hyunk@xilinx.com>
Acked-by: Michal Simek <michal.simek@xilinx.com>
[Quanyang: original patch taken from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v4.19]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>