drivers: net: ethernet: TSN-switch driver initial check-in
The Xilinx TSN Switch is an integrated solution provided with
3-ports, two ports connecting to network through MAC interfaces
and one port connecting to an internal End Point.
TSN switch driver allows processor to control and program the
TSN Switch hardware settings.
This misc driver will be used to provide the APIs for accessing Switch
Core resgisters. Most of the APIs are exposed to userspace by ioctl.

Signed-off-by: Saurabh Sengar <saurabhs@xilinx.com>
Acked-by: Kedareswara rao Appana <appanad@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
[Quanyang: original patch taken from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v4.19]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>