MLK-15061-2: ASoC: fsl_asrc_m2m: fix asrc m2m can't work in imx8
commit  from
https://source.codeaurora.org/external/imx/linux-imx

Power domain need to be enabled when asrc m2m start to work, and
disabled when it stop.
Switch back to use the pm_runtime_get_sync and pm_runtime_put_sync
for which is removed in commit 1a3d82e08fa2 ("ASoC: fsl: refine
the asrc driver for imx8qm").

Signed-off-by: Shengjiu Wang <shengjiu.wang@freescale.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>