MLK-14663-2: ASoC: fsl: imx-wm8960: Kill warning for non-gpr boards
commit  from
https://source.codeaurora.org/external/imx/linux-imx

A side effect of commit 5555277e693a7 ("MLK-13574-1: ASoC: imx-wm8960:
remove the gpr dependency") is that a warning was printed for boards
without gpr. This can be confusing.

imx7d boards do not have a gpr setting, so use imx7d-evk-wm8960
compatible string to avoid printing the warning.

Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>