drm: xilinx: drv: Unset the master flag in release wrapper
The preclose and some other operations are triggered only for
DRIVER_LEGACY flag. The current implementation makes a root
user instance as a fake master and revert it in the preclose
callback, but since it's been deprecated, add a release operation
wrapper and move the logic to the wrapper.

While at it, add a comment regarding the fake master logic.

Reported-by: Madhurkiran Harikrishnan <madhurki@xilinx.com>
Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
[Quanyang: original patch taken from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v4.19]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>