drm: xlnx: xlnx_drv: Remove the dt binding
Remove the dt binding for this logical device. Instead, this module
is transformed into more like a library with new functions,
xlnx_drm_pipeline_init() and xlnx_drm_pipeline_exit(). Those
functions will manage a logical master device based on the dt
binding of caller device, then initialize the drm device, a pipeline
device, with the caller device. This simplifies the pipeline model
by allowing subdrivers to be slave device of the pipeline.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
[Quanyang: original patch taken from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v4.19]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>