ivshmem-net: Fix bogus transition to RESET state
commit  from
https://source.codeaurora.org/external/imx/linux-imx

If we are in READY but the remote is still in INIT, we so far fell back
to RESET which caused the setup to get stuck. Fix this by only
transitioning from READY/RUN to RESET in ivshm_net_state_change if the
remote is in RESET as well.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>
(cherry picked from commit b58915e11eba2643d5c68ea0328823a62c21dc49)
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>