MLK-17631-2 usb: cdns3: gadget: fix the bug the non-ep0 can't work after disconnection
commit  from
https://source.codeaurora.org/external/imx/linux-imx

We add endpoints to ep_match_list when adding gadget module, but
we delete the endpoints from the ep_match_list before set configuration.
When the re-enumeration after the new connection, the ep_match_list
is empty, in that case, the non-ep0s have not configurated, the
transfer on them will be failed.

In this commit, we only delete the endpoints from the list when we
remove the gadget module.

Acked-by: Jun Li <jun.li@nxp.com>
Signed-off-by: Peter Chen <peter.chen@nxp.com>
(cherry picked from commit ad13dd3ffdbd337d38992f283af09785dffeaff8)
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>