bcm2835-dma: Add support for per-channel flags
commit  5096a7ab5c38683995f153619ec9bd2b180b1fd3 from
https://github.com/raspberrypi/linux.git

Add the ability to interpret the high bits of the dreq specifier as
flags to be included in the DMA_CS register. The motivation for this
change is the ability to set the DISDEBUG flag for SD card transfers
to avoid corruption when using the VPU debugger.

Signed-off-by: Phil Elwell <phil@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>