sdk_dpaa: store the skb backpointer in the skb headroom
commit  from
https://source.codeaurora.org/external/imx/linux-imx

The skb backpointer is stored right before the FMan buffer, in order to
avoid overwriting. The memory area storing the backpointer was outside of
the skb. This made it hard to guarantee its size.

This patch changes the layout of the skb at buffer seed time: the area
reserved for storing the skb backpointer is part of the skb's headroom.
This makes it easier to track if the backpointer can be safely stored
when recycling the buffer.

Signed-off-by: Camelia Groza <camelia.groza@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>