MLK-19053 crypto: caam - RNG4 TRNG errata
commit  from
https://source.codeaurora.org/external/imx/linux-imx

 The TRNG as used in RNG4, used in CAAM has a documentation issue. The
effect is that it is possible that the entropy used to instantiate the
DRBG may be old entropy, rather than newly generated entropy. There is
proper programming guidance, but it is not in the documentation.

Signed-off-by: Aymen Sghaier <aymen.sghaier@nxp.com>
Signed-off-by: Vipul Kumar <vipul_kumar@mentor.com>
(cherry picked from commit ea2b30c8171acf7624e3d44276737c3878ca900d)

-ported to RNG initialization in ctrl.c
-changed commit headline

Signed-off-by: Horia Geantă <horia.geanta@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>