staging: mmal-vchiq: Make a mmal_buf struct for passing parameters
commit  59be59efcbd34e215477812fe2875e5994d3423c from
https://github.com/raspberrypi/linux.git

The callback from vchi_mmal to the client was growing lots of extra
parameters. Consolidate them into a single struct instead of
growing the list further.
The struct is associated with the client buffer, therefore there
are various changes to setup various containers for the struct,
and pass the appropriate members.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>