staging: mmal-vchiq: Add support for event callbacks.
commit  51a869790328a5d226ee7ff085781bd8e3f42233 from
https://github.com/raspberrypi/linux.git rpi-5.8.y

(Preparation for the codec driver).
The codec uses the event mechanism to report things such as
resolution changes. It is signalled by the cmd field of the buffer
being non-zero.

Add support for passing this information out to the client.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>