staging: mmal-vchiq: Add support for event callbacks.
commit  19ec81ca2261891e2f5177d350d02286e6049c97 from
https://github.com/raspberrypi/linux.git

(Preparation for the codec driver).
The codec uses the event mechanism to report things such as
resolution changes. It is signalled by the cmd field of the buffer
being non-zero.

Add support for passing this information out to the client.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>