v4l: xilinx-multi-scaler: Check any channel is still running in streamoff
If an application queued buffer while the device is busy in processing
of previous buffer, the application will not call device_run () again.
And at this point if some application running higher channel stops stream,
the device_run () return as there is no buffere for stopping channel.
In this situation the first channel does not get chance to call
device_run() again and will wait for infinite time.

This patch checks if there is/are any continues channels are still running
and if running, start the process again.

Signed-off-by: Suresh Gupta <suresh.gupta@xilinx.com>
Reviewed-by: Satish Kumar Nagireddy <satish.nagireddy.nagireddy@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
[Quanyang: original patch taken from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v4.19]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>