staging: fsl-dpaa2/mac: probe phy as fixed link based on dpmac properties
commit  from
https://source.codeaurora.org/external/imx/linux-imx

Keep in sync the PHY type settings in DPC and Linux device tree.

If the dpmac is connected to a fixed link PHY based on dpc config,
treat it as a fixed-link device, regardless of whether the "phy-handle"
property is present in the device tree node or not.

Signed-off-by: Pankaj Bansal <pankaj.bansal@nxp.com>
Signed-off-by: Ioana Radulescu <ruxandra.radulescu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>