MLK-20095-3: ASoC: fsl: Differentiate between QXP and QM
commit  from
https://source.codeaurora.org/external/imx/linux-imx

On QM the DSP is inside the VPU subsystem while in QXP
it is inside the Audio DMA subsystem. For this reason
there are "subtle" differences.

Introduce new compatible string for QM to help us correctly
configure the DSP depending on the board they run.

dsp_mem_msg structure is shared with the DSP, so by introducing
new member dsp_board_type we can let DSP know on which target it runs.

Reviewed-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>